+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added: Support folding for CUE elements (file header, imports, import groups, structs, lists, interpolations, multiline strings, attributes). 
 - Added: Settings to control the default folding state of CUE elements.
 - Fix [#27](https://github.com/nexantic/intellij-cue/issues/27), file starting with attribute raised as invalid
+- Added: Highlighting setting for attributes.
 
 ## [0.5.0]
 - Support variable escape prefix of the language specification.